refactor and experiment with rsc
